requirements.txt
alembic==0.8.10
appdirs==1.4.0
backports.shutil-get-terminal-size==1.0.0
blinker==1.4
cffi==1.9.1
click==6.7
colorama==0.3.7
Counter==1.0.0
cryptography==1.7.2
decorator==4.0.11
dominate==2.3.1
enum34==1.1.6
et-xmlfile==1.0.1
Flask==0.12
Flask-Login==0.4.0
Flask-Migrate==2.0.3
Flask-MySQLdb==0.2.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
idna==2.2
ipaddress==1.0.18
ipython==5.2.2
ipython-genutils==0.1.0
itsdangerous==0.24
jdcal==1.3
Jinja2==2.9.5
Mako==1.0.6
MarkupSafe==0.23
MySQL-python==1.2.3
mysqlclient==1.3.9
olefile==0.44
openpyxl==2.4.5
packaging==16.8
paramiko==2.1.1
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.13
ptyprocess==0.5.1
pyasn1==0.2.2
pycparser==2.17
Pygments==2.2.0
pyparsing==2.1.10
python-editor==1.0.3
python-memcached==1.58
scandir==1.4
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.5
traitlets==4.3.1
virtualenv==15.1.0
visitor==0.1.3
wcwidth==0.1.7
Werkzeug==0.11.15
win-unicode-console==0.5
WTForms==2.1
